pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["DenisKhudyakov <denis.hudyakov27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
ipython = "^8.26.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
coverage = "^7.6.1"
setuptools = "^72.1.0"


[tool.poetry.group.lint.dependencies]
black = "^24.8.0"
mypy = "^1.11.1"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
